// app/services/types/home.ts
export interface ServicesType {
  id: number;
  name: string;
  icon: string;
  image: string;
  description: string;
  service_categories: {
    category: {
      name: string;
      image: string;
    };
    items: {
      id: number;
      service_category_id: number;
      item_id: number;
      item: {
        name: string;
        description: string;
        unit: string;
        price: number;
        quantity: number;
        status: number;
        created_at: string;
        updated_at: string;
      };
    }[];
  }[];
}